refactor: move query to new 'internals' module
