+++ tests/modes/test_Query.py
@@ -1,4 +1,4 @@
-from ulauncher.api.shared.query import Query
+from ulauncher.internals.query import Query
 
 
 class TestQuery:
