+++ ulauncher/modes/BaseMode.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Iterable
 
 from ulauncher.api.result import Result
-from ulauncher.api.shared.query import Query
+from ulauncher.internals.query import Query
 
 
 class BaseMode:
