+++ ulauncher/modes/extensions/ExtensionMode.py
@@ -4,7 +4,7 @@ import html
 from typing import Any, Generator
 
 from ulauncher.api.result import Result
-from ulauncher.api.shared.query import Query
+from ulauncher.internals.query import Query
 from ulauncher.modes.BaseMode import BaseMode
 from ulauncher.modes.extensions.ExtensionController import ExtensionController
 from ulauncher.modes.extensions.ExtensionSocketServer import ExtensionSocketServer
